% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-utils.R
\name{cached_fetch}
\alias{cached_fetch}
\title{Create cached version of a fetch function}
\usage{
cached_fetch(fetch_func, max_age_hours = 24)
}
\arguments{
\item{fetch_func}{Fetch function to cache}

\item{max_age_hours}{Maximum age of cached results}
}
\description{
Create cached version of a fetch function
}
