% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-calcs.R
\name{calculate_round}
\alias{calculate_round}
\title{Helper function for \code{get_fixture,betting_data}}
\usage{
calculate_round(data_frame)
}
\arguments{
\item{data_frame}{A data frame with match-level data and a Date column}
}
\description{
Work out round number of each game from day and week.
Games from Thursday through Wednesday go in same Round.
}
