% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calculations.R
\name{calculate_season_carryover}
\alias{calculate_season_carryover}
\title{Find new season ELO score by applying carryover factor.}
\usage{
calculate_season_carryover(elo, initial_team = 1500, weight = 0.5)
}
\arguments{
\item{elo}{An ELO rating, typically taken as the end of season value.}

\item{initial_team}{Rating given to the intial team. All values are scaled towards this value}

\item{weight}{A weighting for how much to regress the score towards \code{initial_team}.
A value of 1 would not regress the ELO rating at all
while a value of 0 would regress all the way to \code{initial_team}}
}
\value{
A numeric value indicating the new ELO rating
}
\description{
\code{calculate_season_carryover} returns an ELO rating that is scaled towards the mean based on a carryover weight.
}
\details{
INSERT DESCRIPTION
}
\examples{
calculate_season_carryover(1600, initial_team = 1500, weight = 0.3)
calculate_season_carryover(1400, initial_team = 1550, weight = 0.5)

\dontrun{
calculate_season_carryover(1650, weight = -10)
}
}
