% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-utils.R
\name{clear_fitzroy_cache}
\alias{clear_fitzroy_cache}
\title{Clear all fitzRoy cache}
\usage{
clear_fitzroy_cache(confirm = FALSE)
}
\arguments{
\item{confirm}{If TRUE, skip confirmation prompt}
}
\description{
Clear all fitzRoy cache
}
