% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-helpers.R
\name{convert_results}
\alias{convert_results}
\title{Convert AFL Men's results into long format}
\usage{
convert_results(results)
}
\arguments{
\item{results}{A dataframe that has been returned from get_match_results}
}
\value{
A data frame with match results where each row is a team-match combination
}
\description{
\code{convert_results} returns a dataframe containing the results in long format.
}
\details{
The standard results returned by afltables.com will be in wide format.
This is useful for game based analysis but less so for team based ones. This function converts the data into long format for easier analysis.
}
\examples{
\donttest{
results <- get_match_results()
convert_results(results)
}
}
