% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_afltables}
\alias{dictionary_afltables}
\title{AFL Tables Data Dictionary: AFL Tables Column Formats}
\format{
An object of class \code{data.frame} with 81 rows and 2 columns.
}
\usage{
dictionary_afltables
}
\description{
A dataframe containing the data types and formats of columns for AFL Tables data, used in the \code{check_and_convert()} function to ensure correct column types.
}
\keyword{internal}
