% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ladder.R
\name{fetch_afl_ladder}
\alias{fetch_afl_ladder}
\title{Fetch ladder}
\usage{
fetch_afl_ladder(season = NULL, round_number = NULL, comp = "AFLM")

fetch_afl_ladder(season = NULL, round_number = NULL, comp = "AFLM")
}
\arguments{
\item{season}{season in YYYY format}

\item{comp}{One of "AFLM" or "AFLW"}

\item{source}{One of "AFL" (default), "Footywire", "AFLTables"}

\item{round}{round number}
}
\value{
returns a dataframe with the fixture that matches season, round.

returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Ladder for the relevant Season and Round from the AFL.com.au website.

Returns the Ladder for the relevant Season and Round from the AFL.com.au website.
}
\examples{
\dontrun{
get_ladder(2020, round = 1)
}
\dontrun{
get_afl_ladder(2020, round = 1)
}
}
