% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_awards.R
\name{fetch_awards_allaustralian}
\alias{fetch_awards_allaustralian}
\title{Fetch AFL All-Australian Team or Squad}
\usage{
fetch_awards_allaustralian(season, type = c("team", "squad"))
}
\arguments{
\item{season}{A single year (e.g., 2023)}

\item{type}{Either "team" (final 22) or "squad" (initial 44)}
}
\value{
A tibble with player/team info
}
\description{
Fetch AFL All-Australian Team or Squad
}
