% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_awards.R
\name{fetch_awards_brownlow}
\alias{fetch_awards_brownlow}
\title{Fetch Brownlow Medal Data}
\usage{
fetch_awards_brownlow(season, type = c("player", "team"))
}
\arguments{
\item{season}{A single numeric year (e.g., 2024).}

\item{type}{Either "player" (default) or "team".}
}
\value{
A tibble with Brownlow vote data.
}
\description{
Fetch Brownlow Medal Data
}
