% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-coach-details.R
\name{fetch_coach_details}
\alias{fetch_coach_details}
\title{Fetch coach details}
\usage{
fetch_coach_details(source = "afltables")
}
\arguments{
\item{source}{The source to use. By default this is "afltables".}
}
\value{
A tibble with coach details.
}
\description{
\code{fetch_coach_details} returns coach details such as teams coached,
number of wins, losses and draws, premierships and grand finals.

By default the source used will be AFL Tables.

\code{\link[=fetch_coach_details_afltables]{fetch_coach_details_afltables()}} can be called directly and return data from AFL Tables.
}
\examples{
\dontrun{
coaches <- fetch_coach_details()
}

}
\seealso{
\itemize{
\item \link{fetch_coach_details_afltables} for AFL Tables data.
}
}
\concept{fetch coach functions}
