% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-coaches-votes.R
\name{fetch_coaches_votes}
\alias{fetch_coaches_votes}
\alias{scrape_coaches_votes}
\title{Fetch Coaches Votes}
\usage{
fetch_coaches_votes(
  comp = "AFLM",
  season,
  round = 1:27,
  team = c("Adelaide Crows", "Brisbane Lions", "Carlton", "Collingwood", "Essendon",
    "Fremantle", "Geelong Cats", "Gold Coast Suns", "GWS Giants", "Hawthorn",
    "Melbourne", "North Melbourne", "Port Adelaide", "Richmond", "St Kilda",
    "Sydney Swans", "West Coast Eagles", "Western Bulldogs")
)

scrape_coaches_votes(comp, season, round, finals)
}
\arguments{
\item{comp}{One of "AFLM" (default) or "AFLW"}

\item{season}{Season in YYYY format. This can be an array of seasons.}

\item{round}{Optional - Round number. For finals this is the number of H&A rounds plus the Finals week.}

\item{team}{Optional - Team or teams whose matches should be retrieved (defaults to all teams)}
}
\value{
A data frame with columns: Season, Round, Finals, Home.Team, Away.Team, Player.Name, Coaches.Votes
}
\description{
\code{fetch_coaches_votes} returns all coaches votes for input season/s, round/s, and/or team's matches.
The function calls a core \code{scrape_coaches_votes} function which scrapes the AFLCA website for coaches votes
for a particular season, round and competition.
}
\examples{
\dontrun{
# Return all coaches votes across all seasons
fetch_coaches_votes(comp = "AFLM", season = 2007:2021)
fetch_coaches_votes(comp = "AFLW", season = 2018:2021)

# Return all coaches votes for a particular round
fetch_coaches_votes(comp = "AFLM", season = 2021, round = 24)
fetch_coaches_votes(comp = "AFLW", season = 2021, round = 9)

# Return all coaches votes for a particular team
fetch_coaches_votes(comp = "AFLM", season = 2021, team = "Western Bulldogs")
fetch_coaches_votes(comp = "AFLW", season = 2021, team = "Western Bulldogs)

# Return all coaches votes for a particular match
fetch_coaches_votes(comp = "AFLM", season = 2021, round = 24, team = "Western Bulldogs")
fetch_coaches_votes(comp = "AFLW", season = 2021, round = 9, team = "Western Bulldogs)
}

}
