% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_supercoach_dreamteam.R
\name{fetch_fantasy_scores}
\alias{fetch_fantasy_scores}
\title{Fetch AFL Fantasy (Dream Team) Scores}
\usage{
fetch_fantasy_scores(year = 2025, rounds = 1:30)
}
\arguments{
\item{year}{Integer. AFL season year.}

\item{rounds}{Integer vector. Rounds to fetch (default: 1:30).}
}
\value{
A data frame of Dream Team scores.
}
\description{
Fetch AFL Fantasy (Dream Team) Scores
}
