% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-fixture.R
\name{fetch_fixture_footywire}
\alias{fetch_fixture_footywire}
\title{Get upcoming fixture from https://www.footywire.com}
\usage{
fetch_fixture_footywire(
  season = NULL,
  round_number = NULL,
  convert_date = FALSE
)
}
\arguments{
\item{season}{season in YYYY format, defaults to NULL which returns the year corresponding the \code{Sys.Date()}}

\item{round_number}{round number, defaults to NULL which returns all rounds}

\item{convert_date}{logical, if TRUE, converts date column to date format instead of date time.}
}
\value{
Returns a data frame containing the date, teams and venue of each game
}
\description{
\code{get_fixture} returns a dataframe containing upcoming AFL Men's season fixture.
}
\details{
The dataframe contains the home and away team as well as venue.
}
\examples{
\dontrun{
get_fixture(2018)
}
}
