% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-fixture.R
\name{fetch_fixture_squiggle}
\alias{fetch_fixture_squiggle}
\title{Get Squiggle Fixture}
\usage{
fetch_fixture_squiggle(season = NULL, round_number = NULL)
}
\arguments{
\item{season}{season in YYYY format}

\item{round_number}{round number}
}
\value{
returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Fixture for the relevant Season and Round from the squiggle.com API.
}
\details{
This is essentially a wrapper for \code{fetch_squiggle_data}.

Data returned will contain results of the match if it is complete.
}
\examples{
\dontrun{
fetch_fixture_squiggle(2020, round = 1)
}
}
