% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ladder.R
\name{fetch_ladder_afl}
\alias{fetch_ladder_afl}
\title{Get AFL ladder}
\usage{
fetch_ladder_afl(season = NULL, round_number = NULL, comp = "AFLM")
}
\arguments{
\item{season}{season in YYYY format}

\item{round_number}{round number}

\item{comp}{One of "AFLM" or "AFLW"}
}
\value{
returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Ladder for the relevant Season and Round from the AFL.com.au website.
}
\examples{
\dontrun{
get_afl_ladder(2020, round = 1)
}
}
