% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ladder.R
\name{fetch_ladder_afltables}
\alias{fetch_ladder_afltables}
\title{Recreate the ladder for every or any given round and/or season from the afltables.com websit}
\usage{
fetch_ladder_afltables(
  season = NULL,
  round_number = NULL,
  match_results_df = NULL
)
}
\arguments{
\item{season}{An integer of the season or vector of integers for multiple seasons. If empty, all seasons returned}

\item{round_number}{An integer of the round or vector of integers for multiple rounds. If empty, all rounds returned}

\item{match_results_df}{A dataframe that has been returned from get_match_results. If empty \code{get_match_results} will execute first}
}
\value{
Returns a data frame containing a line for each team's ladder position at each round of a season
}
\description{
\code{fetch_ladder_afltables} returns a dataframe containing the ladder for either all seasons and rounds since 1987, or individual rounds/seasons
}
\details{
The dataframe contains information about the Round, Season, Points For/Against, Ladder Position. It can either take in a data frame created using \code{get_match_results}, or if \code{match_results_df} is unspecified, will extract all games using \code{get_match_results}.
Will only allow selecting rounds of the premiership season, not finals.
}
\examples{
\dontrun{
fetch_ladder_afltables()
fetch_ladder_afltables(match_results_df=get_match_results_df, round_number=23, season=1990:2019)
fetch_ladder_afltables(round_number = 10, season = 2019)
}
}
