% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-ladder.R
\name{fetch_ladder_squiggle}
\alias{fetch_ladder_squiggle}
\title{Get Squiggle ladder}
\usage{
fetch_ladder_squiggle(season = NULL, round_number = NULL)
}
\arguments{
\item{season}{season in YYYY format}

\item{round_number}{round number}
}
\value{
returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Ladder for the relevant Season and Round from the squiggle.com API.
}
\details{
This is essetnially a wrapper for \code{fetch_squiggle_data}.
}
\examples{
\dontrun{
fetch_ladder_squiggle(2020, round = 1)
}
}
