% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-outofcontract.R
\name{fetch_outofcontract_footywire}
\alias{fetch_outofcontract_footywire}
\title{Fetch Out of Contract AFL Players from FootyWire}
\usage{
fetch_outofcontract_footywire(year = 2026)
}
\arguments{
\item{year}{Numeric. Year to fetch (e.g. 2026)}
}
\value{
A tibble
}
\description{
Fetch Out of Contract AFL Players from FootyWire
}
