% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-player-details.R
\name{fetch_player_details}
\alias{fetch_player_details}
\alias{fetch_player_details_afltables}
\alias{fetch_player_details_footywire}
\title{Fetch Player Details}
\usage{
fetch_player_details(team = NULL, comp = "AFLM", source = "afl", ...)

fetch_player_details_afltables(team = NULL)

fetch_player_details_footywire(team, current = TRUE)
}
\arguments{
\item{team}{team the player debuted for, defaults to NULL which returns all data}

\item{comp}{One of "AFLM" (default) or "AFLW"}

\item{source}{One of "AFL" (default), "footywire", "afltables"}

\item{...}{Optional parameters passed onto various functions depending on source.}
}
\value{
A Tibble with the details of the relevant players.
}
\description{
\code{fetch_player_details} returns player details such as date of birth, debut
and other details. The exact details that are returned will depend on which
source is provided.

By default the source used will be the official AFL website.

\code{\link[=fetch_player_details_afltables]{fetch_player_details_afltables()}} and \code{\link[=fetch_player_details_footywire]{fetch_player_details_footywire()}}
can be called directly and return data from AFL Tables and Footywire
respectively.
}
\examples{
\dontrun{
# Return data for current Hawthorn players
fetch_player_details("Hawthorn")

}

}
\seealso{
\itemize{
\item \link{fetch_player_details_afltables} for AFL Tables data.
\item \link{fetch_player_details_footywire} for Footywire data.
}
}
\concept{fetch player details functions}
