% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-results.R
\name{fetch_results}
\alias{fetch_results}
\title{Fetch Results}
\usage{
fetch_results(
  season = NULL,
  round_number = NULL,
  comp = "AFLM",
  source = "AFL",
  ...
)
}
\arguments{
\item{season}{season in YYYY format, defaults to NULL which returns the year corresponding the \code{Sys.Date()}}

\item{round_number}{round number, defaults to NULL which returns all rounds}

\item{comp}{One of "AFLM" (default) or "AFLW"}

\item{source}{One of "AFL" (default), "footywire", "afltables"}

\item{...}{Optional paramters passed onto various functions depending on source}
}
\value{
returns a dataframe with the results that matches season, round.
}
\description{
Returns the Results for the relevant Season and optionally Round from various sources.
}
\examples{
\dontrun{
fetch_results(2020, round = 1)
}
}
