% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-results.R
\name{fetch_results_afl}
\alias{fetch_results_afl}
\title{Fetch AFL.com results}
\usage{
fetch_results_afl(season = NULL, round_number = NULL, comp = "AFLM")
}
\arguments{
\item{season}{season in YYYY format, defaults to NULL which returns the year corresponding the \code{Sys.Date()}}

\item{round_number}{round number, defaults to NULL which returns all rounds}

\item{comp}{One of "AFLM" (default) or "AFLW"}
}
\value{
returns a dataframe with the results that matches season, round.
}
\description{
Returns the Results for the relevant Season and Round from the AFL.com.au website.
}
\examples{
\dontrun{
fetch_results_afl(2020, round = 1)
}
}
