% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-results.R
\name{fetch_results_afltables}
\alias{fetch_results_afltables}
\title{Fetch match results from afltables.com}
\usage{
fetch_results_afltables(season = NULL, round_number = NULL)
}
\arguments{
\item{season}{season in YYYY format, defaults to NULL which returns the year corresponding the \code{Sys.Date()}}

\item{round_number}{round number, defaults to NULL which returns all rounds}
}
\value{
Returns a data frame containing a line for each match
}
\description{
\code{fetch_results_afltables} returns a dataframe containing all match results from 1897-current
}
\details{
The dataframe contains information about the Date, teams involved, scores and venue. It comes from afltables 'big lists' section. This is a limited dataset but is very fast to access.
It generally is updated on the day after the last game
}
\examples{
\dontrun{
fetch_results_afltables()
}
}
