% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-results.R
\name{fetch_results_squiggle}
\alias{fetch_results_squiggle}
\title{Get Squiggle Results}
\usage{
fetch_results_squiggle(season = NULL, round_number = NULL)
}
\arguments{
\item{season}{season in YYYY format}

\item{round_number}{round number}
}
\value{
returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Results for the relevant Season and Round from the squiggle.com API.
}
\details{
This is essentially a wrapper for \code{fetch_squiggle_data}.

Data returned will contain only completed matches. Use \code{fetch_fixture_squiggle} to return non completed matches.
}
\examples{
\dontrun{
fetch_results_squiggle(2020, round = 1)
}
}
