% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_awards.R
\name{fetch_rising_star}
\alias{fetch_rising_star}
\title{Fetch AFL Rising Star Nominations or Stats}
\usage{
fetch_rising_star(
  season,
  round_number = NULL,
  type = c("nominations", "stats")
)
}
\arguments{
\item{season}{Integer. The year of interest (e.g. 2024).}

\item{round_number}{Integer. Optional. If NULL and type = "stats", scrapes all rounds.}

\item{type}{Character. Either "nominations" (default) or "stats".}
}
\value{
A tibble with Rising Star data.
}
\description{
Fetch AFL Rising Star Nominations or Stats
}
\examples{
fetch_rising_star(2024, type = "nominations")
fetch_rising_star(2024, round_number = 5, type = "stats")
fetch_rising_star(2024, type = "stats")
}
