% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_supercoach_dreamteam.R
\name{fetch_supercoach_scores}
\alias{fetch_supercoach_scores}
\title{Fetch Supercoach Scores}
\usage{
fetch_supercoach_scores(year = 2025, rounds = 1:30)
}
\arguments{
\item{year}{Integer. AFL season year.}

\item{rounds}{Integer vector. Rounds to fetch (default: 1:30).}
}
\value{
A data frame of Supercoach scores.
}
\description{
Fetch Supercoach Scores
}
