% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_team_stats.R
\name{fetch_team_stats_footywire}
\alias{fetch_team_stats_footywire}
\title{Fetch Team Statistics from Footywire}
\usage{
fetch_team_stats_footywire(
  season,
  summary_type = c("totals", "averages", "opp_totals", "opp_averages", "diff_totals",
    "diff_averages")
)
}
\arguments{
\item{season}{Integer. The season to fetch.}

\item{summary_type}{One or more of: "totals", "averages", "opp_totals", "opp_averages", "diff_totals", "diff_averages".}
}
\value{
A tibble of team statistics.
}
\description{
Fetch Team Statistics from Footywire
}
\keyword{internal}
