% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_team_stats.R
\name{fetch_team_stats_vflstats}
\alias{fetch_team_stats_vflstats}
\title{Fetch VFLM/VFLW Team Stats from vflstats}
\usage{
fetch_team_stats_vflstats(
  season = 2025,
  summary_type = c("totals", "averages"),
  comp = c("VFLM", "VFLW")
)
}
\arguments{
\item{season}{Integer. A year from 2021 onward.}

\item{summary_type}{Either "totals" or "averages".}

\item{comp}{Competition code: "VFLM" or "VFLW".}
}
\value{
A tibble of team-level statistics.
}
\description{
Fetch VFLM/VFLW Team Stats from vflstats
}
\keyword{internal}
