% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_calculations.R
\name{find_expected_outcome}
\alias{find_expected_outcome}
\title{Find the expected outcome given an ELO difference.}
\usage{
find_expected_outcome(elo_difference, M = 400)
}
\arguments{
\item{elo_difference}{Difference in elo ratings between two teams. Can be a positive or negative number.}

\item{M}{Test}
}
\value{
A numeric value between 0 and 1, where values >0.5 indicate a win
while values of <0.5 indicate a loss.
A value of 0.5 indicates a draw
}
\description{
\code{find_expected_outcome} returns the expected outcome of winning based on an ELO points difference.
}
\details{
This is a generic ELO function in the form of prob = 1/1+10^(elo/M)
where elo is the elo difference between two teams and
M is the scaling factor for the elo difference (EXPAND)
}
\examples{
find_expected_outcome(100, M = 400)
find_expected_outcome(0)

\dontrun{
find_expected_outcome("a")
}
}
