% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitzRoy-package.R
\docType{package}
\name{fitzRoy-package}
\alias{fitzRoy}
\alias{fitzRoy-package}
\title{fitzRoy: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimmyday12/fitzRoy}
  \item Report bugs at \url{https://github.com/jimmyday12/fitzRoy/issues}
}

}
\author{
\strong{Maintainer}: James Day \email{jamesthomasday@gmail.com}

Authors:
\itemize{
  \item Robert Nguyen \email{nguyen.n.robert@gmail.com}
}

}
\keyword{internal}
