% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fixture}
\alias{fixture}
\title{Fixture for upcoming afl seasons.}
\format{A data frame with 198 rows and 7 variables:
\describe{
\item{Date}{Date of game}
\item{Season}{Season of game}
\item{Season.Game}{Game of the season}
\item{Round}{Round of season}
\item{Home.Team}{Designated home team}
\item{Away.Team}{Designated away team}
\item{Venue}{Venue of game}
}}
\source{
\url{http://www.afltables.com}
}
\usage{
fixture
}
\description{
A dataset for games in the upcoming afl season
}
\keyword{datasets}
