% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-helpers.R
\name{footywire_html}
\alias{footywire_html}
\title{Helper function for \code{get_footywire_stats}}
\usage{
footywire_html(x, id)
}
\arguments{
\item{x}{URL of the match}

\item{id}{Match ID number}
}
\value{
A data frame with advanced player results
}
\description{
Helper function for \code{get_footywire_stats}
}
