% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colour_palettes.R
\name{get_afl_colour_palettes}
\alias{get_afl_colour_palettes}
\title{Returns a table with the colour palettes for all teams}
\usage{
get_afl_colour_palettes()
}
\value{
a data table containing team long name, team abbreviation, and colours
}
\description{
\code{get_afl_colour_palettes} returns a data frame containing the AFL team's primary, secondary and tertiary colours as applicable
The data for this function is hosted on github.
}
\examples{
\dontrun{
# Gets all data
get_afl_colour_palettes()
}

}
