% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-afl.R
\name{get_afl_cookie}
\alias{get_afl_cookie}
\title{Get AFL Stats cookie (internal function)}
\usage{
get_afl_cookie()
}
\value{
token code
}
\description{
Gets a cookie from http://www.afl.com.au/ to authenticate
further requests.
}
\examples{
\dontrun{
cookie <- get_afl_cookie()
}
}
