% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_afl_fixture.R
\name{get_afl_fixture}
\alias{get_afl_fixture}
\title{Get AFL fixture}
\usage{
get_afl_fixture(season = NULL, round_number = NULL, comp = "AFLM")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_fixture_afl()} instead
}
\examples{
#
\dontrun{
get_afl_fixture(2020, 1)
# ->
fetch_fixture_afl(2020, 1, "AFLM")
}
}
\keyword{internal}
