% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afltables_player.R
\name{get_afltables_player}
\alias{get_afltables_player}
\title{Return afltables playr match stats}
\usage{
get_afltables_player(match_urls)
}
\value{
data table of afltables.com match results, with a row per player per match.
}
\description{
\code{get_afltables_player} returns a character vector containing match URLs for the specified date range
}
\details{
This function returns the full afltables.com match stats for each player and each game specified in \code{match_urls}.
It is useful to use the helper function \code{get_afltables_urls} to return these or simply navigate to afltables.com
and find the match of interest.
}
