% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_cookie}
\alias{get_aflw_cookie}
\title{Get AFL Stats cookie}
\usage{
get_aflw_cookie()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Replaced with a standard function to return cookies for either mens or womens data
}
\examples{
#
\dontrun{
get_aflw_cookie()
# ->
get_afl_cookie()
}
}
\keyword{internal}
