% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_detailed_data}
\alias{get_aflw_detailed_data}
\title{Get detailed AFLW data}
\usage{
get_aflw_detailed_data(matchids)
}
\arguments{
\item{matchids}{vector of match IDs, like those returned by
\code{get_aflw_match_data()}}
}
\value{
Dataframe with detailed match data. Each row is a match.
}
\description{
Get detailed AFLW data
}
\examples{
\dontrun{
get_aflw_detailed_data(c("CD_M20172640101", "CD_M20172640102"))
}
}
