% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_match_data}
\alias{get_aflw_match_data}
\title{Get AFLW match data}
\usage{
get_aflw_match_data()
}
\value{
a data frame
}
\description{
Retrieves all available AFLW match data.
}
\examples{
get_aflw_match_data()
}
