% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_aflw-player_stats.R
\name{get_aflw_player_stats}
\alias{get_aflw_player_stats}
\title{Return get match stats for all current AFLW matches}
\usage{
get_aflw_player_stats(
  start = 2017,
  end = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_player_stats_fryzigg()} instead
}
\examples{
#
\dontrun{
get_aflw_player_stats(2017, 2018)
# ->
fetch_player_stats_fryzigg(2017:2018, comp = "AFLW")
}
}
\keyword{internal}
