% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_round_data}
\alias{get_aflw_round_data}
\title{Get match data (internal function)}
\usage{
get_aflw_round_data(roundid, cookie)
}
\arguments{
\item{roundid}{a round ID string}

\item{cookie}{a cookie produced by \code{get_womens_cookie()}}
}
\value{
a dataframe containing match data
}
\description{
For a given round ID, get the data for each match played in that round. Use
the column \code{roundId} in the dataframe created by the \code{get_rounds()} function
to specify matches to fetch.
}
\examples{
\dontrun{
get_aflw_round_data("CD_R201826401", get_aflw_cookie())
}
}
