% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_rounds}
\alias{get_aflw_rounds}
\title{Get rounds (internal function)}
\usage{
get_aflw_rounds(cookie)
}
\arguments{
\item{cookie}{a cookie produced by \code{get_aflw_cookie()}}
}
\value{
A dataframe with information about each round
}
\description{
Returns data frame for available round data. Includes the rounds played,
as well as identifiers to make further requests, importantly the roundId.
}
\examples{
\donttest{
get_aflw_rounds(get_aflw_cookie())
}
}
