% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_detailed_data}
\alias{get_detailed_data}
\title{Get detailed AFLW data}
\usage{
get_detailed_data(matchid_vector, cookie)
}
\arguments{
\item{matchid_vector}{vector of match IDs, like those returned by
\code{get_aflw_match_data()}}

\item{cookie}{}
}
\value{
a long data frame with detailed match statistics
}
\description{
Get detailed AFLW data
}
\examples{
get_detailed_data(c("CD_M20172640101", "CD_M20172640102"), 
                           get_aflw_cookie())
}
