% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-calcs.R
\name{get_fixture}
\alias{get_fixture}
\title{Get upcoming fixture from footywire.com}
\usage{
get_fixture(season = 2018)
}
\arguments{
\item{season}{Season to return.}
}
\value{
Returns a data frame containing the date, teams and venue of each game
}
\description{
\code{get_fixture} returns a dataframe containing upcoming fixture
}
\details{
The dataframe contains the home and away team as well as venue.
}
\examples{
\dontrun{
get_fixture(2018)
}
}
