% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_footywire-calcs.R
\name{get_footywire_betting_odds}
\alias{get_footywire_betting_odds}
\title{Get AFL match betting odds from https://www.footywire.com}
\usage{
get_footywire_betting_odds(
  start_season = "2010",
  end_season = lubridate::year(Sys.Date())
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_betting_odds_footywire()} instead
}
\examples{
#
\dontrun{
get_footywire_betting_odds(2017, 2018)
# ->
fetch_betting_odds_footywire(2017, 2018)
}
}
\keyword{internal}
