% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_footywire-results.R
\name{get_footywire_match_ids}
\alias{get_footywire_match_ids}
\title{Get aftables match ids}
\usage{
get_footywire_match_ids(season)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_footywire_match_ids()} instead
}
\examples{
#
\dontrun{
get_footywire_match_ids(2020, 1)
# ->
fetch_footywire_match_ids(2020)
}
}
\keyword{internal}
