% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_footywire-calcs.R
\name{get_footywire_stats}
\alias{get_footywire_stats}
\title{Scrape footywire player statistics.}
\usage{
get_footywire_stats(ids)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_footywire_stats()} instead
}
\examples{
#
\dontrun{
get_footywire_stats(5000)
# ->
fetch_footywire_stats(5000)
}
}
\keyword{internal}
