% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_fryziggafl-player-stats.R
\name{get_fryzigg_stats}
\alias{get_fryzigg_stats}
\title{Return get match stats from fryziggafl.net/api/}
\usage{
get_fryzigg_stats(start = 1897, end = as.numeric(format(Sys.Date(), "\%Y")))
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_player_stats_fryzigg()} instead
}
\examples{
#
\dontrun{
get_fryzigg_stats(2020, 2021)
# ->
fetch_player_stats_fryzigg(2020:2021)
}
}
\keyword{internal}
