% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-afltables.R
\name{get_full_team_afltables}
\alias{get_full_team_afltables}
\title{Get full team name from two-letter abbreviation}
\usage{
get_full_team_afltables(short)
}
\arguments{
\item{short}{Two-letter abbreviation}
}
\value{
Full team name
}
\description{
Get full team name from two-letter abbreviation
}
\keyword{internal}
