% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-helpers.R
\name{get_match_data}
\alias{get_match_data}
\title{Helper function for \code{get_footywire_stats}}
\usage{
get_match_data(id)
}
\arguments{
\item{id}{A match id from afltables}
}
\description{
Helper function for \code{get_footywire_stats}
}
