% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_afltables-basic.R
\name{get_match_results}
\alias{get_match_results}
\title{Get basic match results from afltables.com}
\usage{
get_match_results(season = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_results_afltables()} instead
}
\examples{
#
\dontrun{
get_match_results()
# ->
fetch_results_afltables()
}
}
\keyword{internal}
