% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-squiggle-data.R
\name{get_squiggle_data}
\alias{get_squiggle_data}
\title{Access Squiggle data using the squiggle API service.}
\usage{
get_squiggle_data(
  query = c("teams", "sources", "games", "tips", "ladder", "standings", "virtual", "pav"),
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_squiggle_data()} instead
}
\examples{
#
\dontrun{
get_squiggle_data()
# ->
fetch_squiggle_data()
}
}
\keyword{internal}
