% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_afltables}
\alias{mapping_afltables}
\title{AFL Tables Mapping: AFL Tables Naming Convention Mapping}
\format{
An object of class \code{character} of length 25.
}
\usage{
mapping_afltables
}
\description{
A dataframe used for mapping names of data columns when scraping match statistics using the \code{scrape_afltables_match()} function.
}
\keyword{internal}
