% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{parse_team_abbr}
\alias{parse_team_abbr}
\title{Internal function to ensure names match between different sources and also name changes.
This gets applied to any web scraper}
\usage{
parse_team_abbr(team_name)
}
\arguments{
\item{team_name}{Team name}
}
\description{
Internal function to ensure names match between different sources and also name changes.
This gets applied to any web scraper
}
