% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_mapping_afltables}
\alias{player_mapping_afltables}
\title{AFL Tables player mapping: AFL Tables Player Name Mapping}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13166 rows and 9 columns.
}
\usage{
player_mapping_afltables
}
\description{
A dataframe used for mapping player name and ID columns when scraping match statistics using the \code{scrape_afltables_match()} function.
}
\keyword{internal}
