% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_stats}
\alias{player_stats}
\title{Advanced player match statistics.}
\format{A data frame with 71,104 rows and 42 variables:
\describe{
\item{Date}{date of match, in yyyy-mm-dd format}
\item{Season}{season of match, in yyyy format}
\item{Round}{round of match}
...
}}
\source{
\url{http://www.footywire.com}
}
\usage{
player_stats
}
\description{
A dataset containing the match player statitics from footywire.com for matches since 2010.
}
\keyword{datasets}
