% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score-worm.R
\name{plot_score_worm}
\alias{plot_score_worm}
\title{Return afl match score worm}
\usage{
plot_score_worm(providerId)
}
\arguments{
\item{providerId}{Integer. AFL match ID.}
}
\value{
A ggplot2 plot object showing the AFL match score worm.
}
\description{
\code{plot_score_worm} returns a plot of an AFL match score worm.
}
