% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score-worm.R
\name{plot_score_worm_totals}
\alias{plot_score_worm_totals}
\title{Plot Score Worm Totals}
\usage{
plot_score_worm_totals(providerId)
}
\arguments{
\item{providerId}{AFL match ID can be found using \code{fetch_fixture_afl()}}
}
\value{
A ggplot object showing the total score worm.
}
\description{
This function plots the team totals score worm for AFL games.
}
