% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_brownlow_votes.R
\name{predict_brownlow_votes}
\alias{predict_brownlow_votes}
\title{Predict Brownlow votes using AFL player stats, SC and coaches votes}
\usage{
predict_brownlow_votes(
  train_seasons = 2005:2023,
  predict_season = 2025,
  coach_votes_file = "coaches_votes.RDS",
  type = c("strict", "fractional")
)
}
\arguments{
\item{train_seasons}{Numeric vector. Seasons to train on (e.g. 2005:2023)}

\item{predict_season}{Numeric. Season to predict (e.g. 2025)}

\item{coach_votes_file}{Character. Path to RDS file with coaches votes data.}

\item{type}{Character. Either "strict" or "fractional" to determine the leaderboard method.}
}
\value{
A list with:
\itemize{
\item regression_model: lm object
\item strict_vote_totals: tibble with 3-2-1 allocations
\item fractional_vote_totals: tibble with scaled fractional votes
\item round_leaderboard: tibble round by round leaderboard
}
}
\description{
Predict Brownlow votes using AFL player stats, SC and coaches votes
}
