% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afltables_basic.R
\name{replace_teams}
\alias{replace_teams}
\title{Internal function to ensure names match between different sources and also name changes.
This gets applied to any web scraper}
\usage{
replace_teams(team)
}
\arguments{
\item{team}{Team name}
}
\description{
Internal function to ensure names match between different sources and also name changes.
This gets applied to any web scraper
}
