% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_weather}
\alias{results_weather}
\title{Results with weather data.}
\format{A data frame with 15200 rows and 19 variables:
\describe{
\item{date}{Date of game}
}}
\source{
\url{http://www.afltables.com}
}
\usage{
results_weather
}
\description{
A dataset with match results combined with weather data.
}
\keyword{datasets}
