% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{score_progression_raw}
\alias{score_progression_raw}
\title{Scoring progression data from afltables.}
\format{A data frame with 110807 rows and 5 variables:
\describe{
\item{Scoring progression}{x}
}}
\source{
\url{http://www.afltables.com}
}
\usage{
score_progression_raw
}
\description{
A dataset with score progression data from afltables. The data is unprocessed and so is in the simple raw, event format
Data is only available for 2010-2017
}
\keyword{datasets}
