% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-afl.R
\name{team_abr_afl}
\alias{team_abr_afl}
\title{Internal function to return team name abbreviation for AFL API}
\usage{
team_abr_afl(team)
}
\arguments{
\item{team}{Team name}
}
\description{
Internal function to return team name abbreviation for AFL API
}
