% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-afl.R
\name{team_abr_afl2}
\alias{team_abr_afl2}
\title{Internal function to return team name abbreviation for AFL API}
\usage{
team_abr_afl2(team, comp = "AFLM")
}
\arguments{
\item{team}{Team name}

\item{comp}{Competition}
}
\description{
Internal function to return team name abbreviation for AFL API
}
\keyword{internal}
