% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_main.R
\name{update_elo}
\alias{update_elo}
\title{Find a team's new ELO given their previous ELO, home ground advantage and actual result.}
\usage{
update_elo(elo_home, elo_away, margin, returns = c("home", "away", "both"),
  HGA = 10, M, B)
}
\arguments{
\item{elo_home}{The absolute ELO rating of the home team}

\item{elo_away}{The absolute ELO rating of the away team}

\item{margin}{The actual margin of the match, in points, from the home team perspective. Positive numbers indicate a home team win, negative numbers indicate an away team win. A value of 0 indicates a draw.}

\item{returns}{The value to return as an output. Can be either the home team ELO, the away team ELO or both. Returning both will return a vector of two values with the home team ELO listed first.}

\item{HGA}{Home ground advantage}
}
\value{
Returns the ELO of the team specified in \code{returns}
}
\description{
\code{update_elo} returns a new ELO value for the specified team.
}
\details{
This calculates an updated ELO based on the ELO rating of each team,
the home ground advatnage assigned and the actual margin. It can return either the home team ELO, the away team ELO or both
}
\examples{
update_elo(1650, elo_away = 1500, margin = 26, returns = "home", HGA = 20)

}
