% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-calcs.R
\name{update_footywire_stats}
\alias{update_footywire_stats}
\title{Update the included footywire stats data to the specified date.}
\usage{
update_footywire_stats(end_date = Sys.Date())
}
\arguments{
\item{end_date}{X}
}
\value{
Returns a data frame containing player match stats for each match ID
}
\description{
\code{update_footywire_stats} returns a dataframe containing player match stats from footywire.
}
\details{
The dataframe contains both basic and advanced player statistics from each match from 2010 to the specified end date.
}
\examples{
update_footywire_stats()

}
