% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{verify_year}
\alias{verify_year}
\title{Returns year as numeric after verifying the validity of the year.}
\usage{
verify_year(year)
}
\arguments{
\item{year}{haracter string or numeric for year, in "YYYY"ormat}
}
\value{
eturns year as numeric
}
\description{
\code{verify_year} returns a numeric year, after to doing a validity check
}
\examples{
#
\dontrun{
verify_year(2012)
}
}
