% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ladder.R
\name{fetch_ladder}
\alias{fetch_ladder}
\title{Fetch ladder}
\usage{
fetch_ladder(season = NULL, round_number = NULL, comp = "AFLM", source = "AFL")
}
\arguments{
\item{season}{season in YYYY format}

\item{round_number}{round number}

\item{comp}{One of "AFLM" or "AFLW"}

\item{source}{One of "AFL" (default), "Footywire", "AFLTables"}
}
\value{
returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Ladder for the relevant Season and Round from the AFL.com.au website.
}
\examples{
\dontrun{
get_ladder(2020, round = 1)
}
}
