% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitzRoy-package.R
\docType{package}
\name{fitzRoy-package}
\alias{fitzRoy}
\alias{fitzRoy-package}
\title{fitzRoy: Easily Scrape and Process AFL Data}
\description{
An easy to use tool for scraping and processing AFL
    data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimmyday12/fitzRoy}
  \item Report bugs at \url{https://github.com/jimmyday12/fitzRoy/issues}
}

}
\author{
\strong{Maintainer}: James Day \email{jamesthomasday@gmail.com}

Authors:
\itemize{
  \item Robert Nguyen \email{nguyen.n.robert@gmail.com}
  \item Oscar Lane \email{lane.oscar@gmail.com}
}

Other contributors:
\itemize{
  \item Matthew Erbs \email{matthew.erbs@merbla.com} [contributor]
}

}
\keyword{internal}
