% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitzRoy-package.R
\docType{package}
\name{fitzRoy-package}
\alias{fitzRoy}
\alias{fitzRoy-package}
\title{fitzRoy: Easily Scrape and Process AFL Data}
\description{
An easy package for scraping and processing Australia Rules Football (AFL)
    data. 'fitzRoy' provides a range of functions for accessing publicly available data 
    from 'AFL Tables' <https://afltables.com>, 'Footy Wire' <https://www.footywire.com> and
    'The Squiggle' <https://squiggle.com.au>. Further functions allow for easy processing, 
    cleaning and transformation of this data into formats that can be used for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimmyday12/fitzRoy}
  \item Report bugs at \url{https://github.com/jimmyday12/fitzRoy/issues}
}

}
\author{
\strong{Maintainer}: James Day \email{jamesthomasday@gmail.com}

Authors:
\itemize{
  \item Robert Nguyen \email{nguyen.n.robert@gmail.com}
  \item Oscar Lane \email{lane.oscar@gmail.com}
}

Other contributors:
\itemize{
  \item Matthew Erbs \email{matthew.erbs@merbla.com} [contributor]
  \item Jason Zivkovic \email{jaseziv83@gmail.com} [contributor]
}

}
\keyword{internal}
