% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_fixture.R
\name{get_afl_fixture}
\alias{get_afl_fixture}
\title{Get AFL fixture}
\usage{
get_afl_fixture(season = NULL, round_number = NULL, comp = "AFLM")
}
\arguments{
\item{season}{season in YYYY format}

\item{round_number}{round number}

\item{comp}{One of "AFLM" or "AFLW"}
}
\value{
returns a dataframe with the fixture that matches season, round.
}
\description{
Returns the Fixture for the relevent Season and Round from the AFL.com.au website.
}
\examples{
\dontrun{
get_afl_fixture(2020, round = 1)
}
}
