% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-afltables-player.R
\name{get_afltables_stats}
\alias{get_afltables_stats}
\title{Return afltables match stats}
\usage{
get_afltables_stats(start_date = "1897-01-01", end_date = Sys.Date())
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

#' All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_player_stats_afltables()} instead
}
\examples{
#
\dontrun{
get_afltables_stats()
# ->
fetch_player_stats_afltables()
}
}
\keyword{internal}
