% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_cookie}
\alias{get_aflw_cookie}
\title{Get AFL Stats cookie (internal function)}
\usage{
get_aflw_cookie()
}
\value{
token code
}
\description{
Gets a cookie from http://www.afl.com.au/womens/matches/stats to authenticate
further requests.
}
\examples{
cookie <- get_aflw_cookie()
}
