% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_detailed_match_data}
\alias{get_aflw_detailed_match_data}
\title{Get detailed womens match data (internal function)}
\usage{
get_aflw_detailed_match_data(matchid, roundid, competitionid, cookie)
}
\arguments{
\item{matchid}{matchid from \code{get_match_data()}}

\item{roundid}{roundid from \code{get_match_data()}}

\item{competitionid}{competitionid from \code{get_match_data()}}

\item{cookie}{cookie from \code{get_womens_cookie()}}
}
\value{
Dataframe with detailed match data (wide)
}
\description{
Gets detailed match data for a given match. Requires the match, round, and
competition IDs, which are given in the tables produced by
\code{get_aflw_round_data()}
}
\examples{
\dontrun{
get_aflw_detailed_match_data(
  "CD_M20172640101",
  "CD_R201726401", "CD_S2017264", get_aflw_cookie()
)
}
}
