% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_match_data}
\alias{get_aflw_match_data}
\title{Get match data}
\usage{
get_aflw_match_data(roundid, cookie)
}
\arguments{
\item{cookie}{a cookie produced by \code{get_womens_cookie()}}

\item{x}{a dataframe of round metadata, as produces by
\code{process_round_metadata}}
}
\value{
a dataframe containing match data
}
\description{
For a given round ID, get the data for each match played. Use the column
\code{roundId} in the dataframe created by the \code{get_rounds} function to specify
matches to fetch
}
\examples{
get_aflw_match_data("CD_R201826401", get_womens_cookie())
}
