% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_match_data}
\alias{get_aflw_match_data}
\title{Get AFLW match data}
\usage{
get_aflw_match_data(start_year = 2017)
}
\arguments{
\item{start_year}{optional, integer for start year to return match data
onwards from}
}
\value{
a data frame of data for all available AFLW matches
}
\description{
Retrieves AFLW match data for all available matches. Sources data from
\url{https://www.womens.afl/}
}
\examples{
\dontrun{
# All data
get_aflw_match_data()

# 2018 data onward
get_aflw_match_data(start_year = 2018)
}
}
