% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womens_stats.R
\name{get_aflw_match_data}
\alias{get_aflw_match_data}
\title{Get AFLW match data}
\usage{
get_aflw_match_data(start_year = 2017)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_fixture_afl()} instead
}
\examples{
#
\dontrun{
get_aflw_match_data(2020)
# ->
fetch_results_afl(2020, comp = "AFLW")
}
}
\keyword{internal}
