% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-calcs.R
\name{get_fixture}
\alias{get_fixture}
\title{Get upcoming fixture from https://www.footywire.com}
\usage{
get_fixture(season = lubridate::year(Sys.Date()), convert_date = FALSE)
}
\arguments{
\item{season}{Season to return, in yyyy format}

\item{convert_date}{logical, if TRUE, converts date column to date format instead of date time.}
}
\value{
Returns a data frame containing the date, teams and venue of each game
}
\description{
\code{get_fixture} returns a dataframe containing upcoming AFL Men's season fixture.
}
\details{
The dataframe contains the home and away team as well as venue.
}
\examples{
\dontrun{
get_fixture(2018)
}
}
