% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_footywire-calcs.R
\name{get_fixture}
\alias{get_fixture}
\title{Get upcoming fixture from https://www.footywire.com}
\usage{
get_fixture(season = lubridate::year(Sys.Date()), convert_date = FALSE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_fixture_footywire()} instead
}
\examples{
#
\dontrun{
get_fixture(2020)
# ->
fetch_fixture_footywire(2020)
}
}
\keyword{internal}
