% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_footywire-results.R
\name{get_footywire_match_results}
\alias{get_footywire_match_results}
\title{Get footywire Match Results}
\usage{
get_footywire_match_results(season, last_n_matches = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_results_footywire()} instead
}
\examples{
#
\dontrun{
get_footywire_match_results(2020, 1)
# ->
fetch_results_footywire(2020, 1)
}
}
\keyword{internal}
