% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-calcs.R
\name{get_footywire_stats}
\alias{get_footywire_stats}
\title{Scrape footywire player statistics.}
\usage{
get_footywire_stats(ids)
}
\arguments{
\item{ids}{A vector containing match id's to return. Can be a single value or vector of values.}
}
\value{
Returns a data frame containing player match stats for each match ID
}
\description{
\code{get_footywire_stats} returns a dataframe containing player match stats from footywire from 2010 onwards.
}
\details{
The dataframe contains both basic and advanced player statistics from each match specified in the match_id input.
To find match ID, find the relevant matches on https://wwww.footywire.com
}
\examples{
\dontrun{
get_footywire_stats(ids = 5000:5100)
}
}
