% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score-progression.R
\name{get_score_progression_raw}
\alias{get_score_progression_raw}
\title{Get raw score progression data}
\usage{
get_score_progression_raw()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated due to its inefficiency
}
\examples{
#
\dontrun{
get_match_results()
# ->
fetch_results_afltables()
}
}
