% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_progression.R
\name{get_score_progression_raw}
\alias{get_score_progression_raw}
\title{Get raw score progression data}
\usage{
get_score_progression_raw()
}
\value{
Returns a data frame containing raw score progression data
}
\description{
\code{get_score_progression_raw} returns a dataframe raw, unprocessed scoring progression data from afltables.
}
\details{
The data is unprocessed and unstructured but is a starting point for analysis. It only exists for 2010 to 2017.
}
\examples{
\dontrun{
get_score_progession_raw()
}
}
