% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_results}
\alias{match_results}
\title{Resulsts from all AFL Men's matches}
\format{A data frame with 15200 rows and 16 variables:
\describe{
\item{Date}{Date of game}
}}
\source{
\url{http://www.afltables.com}
}
\usage{
match_results
}
\description{
A dataset for all AFL Men's matches since 1897
}
\keyword{datasets}
