% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_results}
\alias{match_results}
\title{Resulsts from all AFL Men's matches}
\format{A data frame with 15200 rows and 16 variables:
\describe{
\item{Game}{todo}
\item{Date}{todo}
\item{Round}{todo}
\item{Home.Team}{todo}
\item{Home.Goals}{todo}
\item{Home.Behinds}{todo}
\item{Home.Points}{todo}
\item{Away.Team}{todo}
\item{Away.Goals}{todo}
\item{Away.Behinds}{todo}
\item{Away.Points}{todo}
\item{Venue}{todo}
\item{Margin}{todo}
\item{Season}{todo}
\item{Round.Type}{todo}
\item{Round.Number}{todo}

}}
\source{
\url{http://www.afltables.com}
}
\usage{
match_results
}
\description{
A dataset for all AFL Men's matches since 1897
}
\keyword{datasets}
