% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_stats}
\alias{player_stats}
\title{Advanced player match statistics.}
\format{An object of class \code{data.frame} with 86460 rows and 43 columns.}
\source{
\url{http://www.footywire.com}
}
\usage{
player_stats
}
\description{
A dataset containing the match player statitics from footywire.com for matches since 2010.
}
\keyword{datasets}
