% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_stats}
\alias{player_stats}
\title{Advanced player match statistics.}
\format{A data frame with 71,104 rows and 42 variables:
\describe{
\item{Date}{date of match, in yyyy-mm-dd format}
\item{Season}{season of match, in yyyy format}
\item{Round}{round of match}
\item{AF}{todo}
\item{B}{todo}
\item{BO}{todo}
\item{CCL}{todo}
\item{CG}{todo}
\item{CL}{todo}
\item{CM}{todo}
\item{D}{todo}
\item{DE}{todo}
\item{ED}{todo}
\item{FA}{todo}
\item{FF}{todo}
\item{G}{todo}
\item{GA}{todo}
\item{GA1}{todo}
\item{HB}{todo}
\item{HO}{todo}
\item{I50}{todo}
\item{ITC}{todo}
\item{K}{todo}
\item{M}{todo}
\item{MG}{todo}
\item{MI5}{todo}
\item{Match_id}{todo}
\item{One.Percenters}{todo}
\item{Opposition}{todo}
\item{Player}{todo}
\item{R50}{todo}
\item{SC}{todo}
\item{SCL}{todo}
\item{SI}{todo}
\item{Status}{todo}
\item{T}{todo}
\item{T5}{todo}
\item{TO}{todo}
\item{TOG}{todo}
\item{Team}{todo}
\item{UP}{todo}
\item{Venue}{todo}
...
}}
\source{
\url{http://www.footywire.com}
}
\usage{
player_stats
}
\description{
A dataset containing the match player statitics from footywire.com for matches since 2010.
}
\keyword{datasets}
