% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score-worm.R
\name{plot_score_worm}
\alias{plot_score_worm}
\title{Plot Score Worm}
\usage{
plot_score_worm(match_id)
}
\arguments{
\item{match_id}{Champion Data match_id (providerId) of the form CD_MSSSS014RRMM where SSSS is the Season, RR is the Round and MM is the Match. e.g. ‘CD_M20240142004' - can be found using \code{fetch_fixture_afl()}}
}
\value{
A ggplot object showing the score worm.
}
\description{
This function plots the score difference score worms for AFL games.
}
