% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score-worm.R
\name{plot_score_worm_totals}
\alias{plot_score_worm_totals}
\title{Return afl match score worm}
\usage{
plot_score_worm_totals(providerId)
}
\arguments{
\item{providerId}{Integer. AFL match ID.}
}
\value{
A ggplot2 plot object showing the total scores of an AFL match.

A ggplot2 plot object showing the total scores of an AFL match.
}
\description{
\code{plot_score_worm_totals} returns a plot of an AFL match score worm showing total scores.
}
