% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-afltables.R
\name{replace_venues}
\alias{replace_venues}
\title{Internal function to ensure venue names match between different sources
and also name changes across time. This gets applied to any web scraper,
transforming all of them to AFL Tables naming conventions.}
\usage{
replace_venues(venue)
}
\arguments{
\item{venue}{Venue name}
}
\description{
Internal function to ensure venue names match between different sources
and also name changes across time. This gets applied to any web scraper,
transforming all of them to AFL Tables naming conventions.
}
