% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_footywire-calcs.R
\name{update_footywire_stats}
\alias{update_footywire_stats}
\title{Update the included footywire stats data to the specified date.}
\usage{
update_footywire_stats(check_existing = TRUE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All \code{get_} functions were replaced with \verb{fetch_*} functions.
Please use \code{fetch_player_stats_footywire()} instead
}
\examples{
#
\dontrun{
update_footywire_stats()
# ->
fetch_player_stats_footywire(2010:2018)
}
}
\keyword{internal}
