% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footywire-calcs.R
\name{update_footywire_stats}
\alias{update_footywire_stats}
\title{Update the included footywire stats data to the specified date.}
\usage{
update_footywire_stats(check_existing = TRUE)
}
\arguments{
\item{check_existing}{A logical specifying if we should check against existing dataset. Defaults to TRUE. Making it false will download all data from all history which will take some time.}
}
\value{
Returns a data frame containing player match stats for each match ID
}
\description{
\code{update_footywire_stats} returns a dataframe containing player match stats from \href{https://www.footywire.com}{footywire}
}
\details{
The dataframe contains both basic and advanced player statistics from each match from 2010 to the specified end date.

This function utilised the included ID's dataset to map known ID's. It looks for any new data that isn't already loaded and proceeds to download it.
}
\examples{
\donttest{
update_footywire_stats()
}
}
